backend-flask/requirements.txt
flask
flask-cors
Flask-AWSCognito

psycopg[binary]
psycopg[pool]

aws-xray-sdk
rollbar
blinker

watchtower

opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp-proto-http 
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-requests